Fix static method handling
The argument count (used to group methods) was not taking into account the implicit `this` parameter in non-static methods.
This meant that, e.g. `public void foo(int)` and `public static void foo(int, int)` would be grouped separately, even though
they both take 2 args, resulting in an error: "java.lang.RuntimeException: Can't have 2 overloads with same arity".

I now handle the implicit `this` parameter in the `parameter-count` and `parameter-types` functions, and removed the special
handling from `wrapper-tail`/`wrapper-multi-tail`, as there can be several overloads of the same arity some of which may be
static and some not, so handling `this` here doesn't really make sense.

I added a test for this case, and removed the previous test as it is now redundant with the test.check tests.