Add ability to filter members
The `filter-fn` option must resolve to a function at macro-expansion time,
since it will be used to determine which members have Clojure wrappers generated.

Also add a filter option to the CLI to filter methods using a regex on a string
representation of the methods, along with an option to print these strings instead
of the wrapped functions.