Refactor to keep member info available at runtime
When there are multiple matching overloads, we need to select the best match at runtime.
We can't just emit the Method/Constructor object into the generated code as there are
no literals for Java objects.
Neither can we realistically look up the object at runtime, as one of the main points
of tortilla is to avoid runtime reflection.
Therefore I've refactored the code to convert the Java objects to MemberInfo records.
These are both nicer to work with a compile time, and also can be made available at runtime
by emitting a call to the record constructor function.