Pass MemberInfo record through to run time
This will allow it to be used to select the best overload at runtime when there is more than one match.
For now the only improvement over the previous behaviour is to prefer a non-vararg overload is one is available
but otherwise it just picks an arbitrary one as before.