Add initial coercer implementation
Currently it will coerce:
- Long to Int so you can use e.g. `(f 1)` instead of `(f (int 1))` in functions that take Integers.
- Double to Float so you can use e.g. `(f 1.0)` instead of `(f (float 1.0))` in functions that take Floats.
- keywords to Enums, so you can use e.g. `(f :VALUE)` instead of `(f EnumType/VALUE)` in functions that take enums.
- vectors to arrays, so you can use e.g. `(f [nil 1 2])` instead of `(f (into-array Long [nil 1 2]))` in functions that take arrays.
- functions into Java functional interface implementations, so you can use clojure functions directly in functions that take
  functional interfaces (i.e. those would accept a lambda expression in Java 8+).
This list is liable to be changed (mostly added to, but things may also be removed if not deemed useful).

Currently, the functional interface support only works with a finite list of functional interfaces known at compile time
(hopefully most, if not all, present in Java itself). Support for others can be easily added by the user with
`(tortilla.coerce/coerce-fn-impl some.ns.OtherFunctionalInterface)`.
This may be improved in the future by automatically adding support when a `@FunctionalInterface` annotation is seen during
wrapper generation.